+++ CHANGES.rst
@@ -37,6 +37,8 @@ Changelog
 - Habilita a busca de objetos de tipo ``sc.embedder``.
   [hvelarde]
 
+- Adiciona um configlet para gerenciar os links no portal tabs.
+  [claytonc.sousa]
 
 2.0a2 (2018-01-11)
 ^^^^^^^^^^^^^^^^^^